requirements.txt
PyQt6>=6.5.0
pyqtgraph>=0.13.0
alpaca-py>=0.13.0
yfinance>=0.2.36
pandas>=2.0.0
numpy>=1.24.0
requests>=2.31.0
